Fixed Multiple bugs and made too many changes to even remember :(

+++ Beta_final/CommandLine_Handler.py
@@ -28,7 +28,8 @@ def parse_arguments(arguments):
     return arg_dict
 
 
-def handle_file_input(input_file, template_path, template_map_path, formats, directories, input_map_path=None):
+def handle_file_input(input_file, template_path, template_map_path, formats, directories, input_map_path=None,
+                      move_ip_map=True):
     # TODO: Implement algorithm to handle file as input
     if(input_map_path==None or input_map_path==''):
         input_path = pathlib2.Path(input_file)
@@ -41,29 +42,50 @@ def handle_file_input(input_file, template_path, template_map_path, formats, dir
         with open(input_map_path, "w") as data_file:
             json.dump(data, data_file, indent=2)
 
-    loaded_data_path = Data_Loader.transfer_values(input_file, template_path, input_map_path, template_map_path,
-                                                   directories['outputs']['loaded'])
+    loaded_data_path = Data_Loader.transfer_values(input_file, template_path, input_map_path, template_map_path)
     row_list = Normalizer.normalize(loaded_data_path, template_map_path, formats)
     normalized_path = Normalizer.writerows(loaded_data_path, row_list, template_map_path)
 
     loaded_data_path_obj = pathlib2.Path(loaded_data_path)
     loaded_data_parent = loaded_data_path_obj.parent
     loaded_data_destination =(loaded_data_parent.joinpath(directories["outputs"])).joinpath(directories['loaded'])
-    print(loaded_data_path)
-    print(loaded_data_destination)
-    shutil.move(loaded_data_path, str(loaded_data_destination))
+    if((loaded_data_destination.joinpath(loaded_data_path_obj.name)).exists()):
+        print("A file with name {} already exists in the destination path provided to move."
+              .format(loaded_data_path_obj.name))
+        response=raw_input("Would you like to replace the file with the newly processed one? (Y/N)")
+        if response.lower() == 'y':
+            os.remove(str(loaded_data_destination.joinpath(loaded_data_path_obj.name)))
+            shutil.move(loaded_data_path, str(loaded_data_destination))
+    else:
+        shutil.move(loaded_data_path, str(loaded_data_destination))
 
     normalized_path_obj = pathlib2.Path(normalized_path)
     normalized_path_parent = normalized_path_obj.parent
     normalized_data_destination = (normalized_path_parent.joinpath(directories["outputs"])).\
         joinpath(directories['normalized'])
-    shutil.move(normalized_path, str(normalized_data_destination))
+    if((normalized_data_destination.joinpath(normalized_path_obj.name)).exists()):
+        print("A file with name {} already exists in the destination path provided to move."
+              .format(normalized_path_obj.name))
+        response=raw_input("Would you like to replace the file with the newly processed one? (Y/N)")
+        if response.lower() == 'y':
+            os.remove(str(normalized_data_destination.joinpath(normalized_path_obj.name)))
+            shutil.move(normalized_path, str(normalized_data_destination))
+    else:
+        shutil.move(normalized_path, str(normalized_data_destination))
 
-    ip_map = pathlib2.Path(input_map_path)
-    ip_parent = ip_map.parent
-    json_templates_path = ip_parent.joinpath(directories['json_maps'])
-
-    shutil.move(str(ip_map), str(json_templates_path))
+    if move_ip_map:
+        ip_map = pathlib2.Path(input_map_path)
+        ip_parent = ip_map.parent
+        json_templates_path = ip_parent.joinpath(directories['json_maps'])
+        if((json_templates_path.joinpath(ip_map.name)).exists()):
+            print("A file with name {} already exists in the destination path provided to move, move not possible."
+                  .format(ip_map.name))
+            response = raw_input("Would you like to replace the file with the newly processed one? (Y/N)")
+            if response.lower() == 'y':
+                os.remove(str(json_templates_path.joinpath(ip_map.name)))
+                shutil.move(str(ip_map), str(json_templates_path))
+        else:
+            shutil.move(str(ip_map), str(json_templates_path))
 
     return
 
@@ -87,43 +109,37 @@ def handle_dir_input(dir_path, template_path, template_map_path, formats, direct
                 json.dump(data, data_file, indent=2)
 
         for file in files_list:
-            handle_file_input(file, template_path, template_map_path, formats, input_map_path)
+            handle_file_input(file, template_path, template_map_path, formats, input_map_path, move_ip_map=False)
         ip_map = pathlib2.Path(input_map_path)
         ip_parent = ip_map.parent
         json_templates_path = ip_parent.joinpath(directories['json_maps'])
-        shutil.move(str(ip_map), str(json_templates_path))
-    else:
-        count=1
-        for file in files_list:
-            if(input_map_path is None) or input_map_path == '':
-                print("Fill in data location information for {}".format(pathlib2.Path(file).name))
-                handle_file_input(file, template_path, template_map_path, formats, directories)
-            # TODO: Remove this unnecessary parts of the code.
-            else:
-                if(count==1):
-                    handle_file_input(file, template_path, template_map_path, formats, directories, input_map_path)
-            count+=1
+        if (json_templates_path.joinpath(ip_map.name)).exists():
+            print("A file with name {} already exists in the destination path provided to move."
+                  .format(ip_map.name))
+            response = raw_input("Would you like to replace the file with the newly processed one? (Y/N)")
+            if response.lower() == 'y':
+                os.remove(str(json_templates_path.joinpath(ip_map.name)))
+                shutil.move(str(ip_map), str(json_templates_path))
+        else:
+            shutil.move(str(ip_map), str(json_templates_path))
 
+    else:
+        for file in files_list:
+            print("Fill in data location information for {}".format(pathlib2.Path(file).name))
+            handle_file_input(file, template_path, template_map_path, formats, directories)
     return
 
 
 def main():
     Config = configparser.ConfigParser()
     Config.read('CommandLine_Config.ini')
-    """outputs_dir = Config.get('DIRECTORIES', 'OUTPUTS_DIR')
-    loaded_dir = Config.get('DIRECTORIES', 'LOADED_DIR')
-    normalized_dir = Config.get('DIRECTORIES', 'NORMALIZED_DIR')
-    json_maps_dir = Config.get('DIRECTORIES', 'INPUT_JSON_TEMPLATES')"""
-    dirs = {'outputs': 'Outputs', 'loaded': 'Loaded', 'noarmalized': 'Normalized', 'json_maps': 'Json_Maps'}
-    """
-    dirs['outputs'] = outputs_dir
-    dirs['loaded'] = loaded_dir
-    dirs['normalized'] = normalized_dir
-    dirs['json_maps'] = json_maps_dir"""
+
+    dirs = {'outputs': 'Outputs', 'loaded': 'Loaded', 'normalized': 'Normalized', 'json_maps': 'Json_Maps'}
 
     datetime_format = Config.get('NORMALIZE_FORMATS', 'datetime')
     date_format = Config.get('NORMALIZE_FORMATS', 'date')
     time_format = Config.get('NORMALIZE_FORMATS', 'time')
+
     formats = {}
     formats["datetime_format"] = datetime_format
     formats["date_format"] = date_format
@@ -152,12 +168,14 @@ def main():
         normalized_dir = outputs_dir.joinpath(dirs['loaded'])
         loaded_dir = outputs_dir.joinpath(dirs["normalized"])
 
+
         if(not(outputs_dir.exists())):
             os.mkdir(str(outputs_dir))
         if(not(normalized_dir.exists())):
             os.mkdir(str(normalized_dir))
         if(not(loaded_dir.exists())):
             os.mkdir(str(loaded_dir))
+
 
         template_map_path = str(template_parent_dir.joinpath(template_fname_split[0] + '.json'))
 
@@ -173,8 +191,15 @@ def main():
 
     if(input_fpath.exists()):
         if(input_fpath.is_file()):
+            input_dir_obj = input_fpath.parent
+            json_maps_dir = input_dir_obj.joinpath(dirs["json_maps"])
+            if (not (json_maps_dir.exists())):
+                os.mkdir(str(json_maps_dir))
             handle_file_input(input_file_or_dir, template_file_path, template_map_path, formats, input_map_path)
         elif(input_fpath.is_dir()):
+            json_maps_dir = input_fpath.joinpath(dirs["json_maps"])
+            if (not (json_maps_dir.exists())):
+                os.mkdir(str(json_maps_dir))
             same_file_formats = False
             file_fmt_response = None
             while (file_fmt_response is None) or file_fmt_response == '':
